+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/analytics/DefaultEventAnalyticsController.kt
@@ -54,6 +54,10 @@ class DefaultEventAnalyticsController(
         }
 
         eventAttributes[EventAttributeName.appState] = appStateMonitor.appState.description
+        appStateMonitor.getBatteryLevel()?.let {
+            eventAttributes[EventAttributeName.batteryLevel] = it
+        }
+        eventAttributes[EventAttributeName.batteryState] = appStateMonitor.getBatteryState().description
 
         eventReporter?.report(SDKEvent(name, eventAttributes))
 
@@ -67,8 +71,14 @@ class DefaultEventAnalyticsController(
     override fun pushHistory(historyEventName: MeetingHistoryEventName) {
         val currentTimeMs = Calendar.getInstance().timeInMillis
         val eventAttributes = mutableMapOf(
-            EventAttributeName.timestampMs to currentTimeMs
+            EventAttributeName.timestampMs to currentTimeMs,
+            EventAttributeName.appState to appStateMonitor.appState.description,
+            EventAttributeName.batteryState to appStateMonitor.getBatteryState().description
         ) as EventAttributes
+
+        appStateMonitor.getBatteryLevel()?.let {
+            eventAttributes[EventAttributeName.batteryLevel] = it
+        }
 
         eventReporter?.report(SDKEvent(historyEventName, eventAttributes))